Status

+++ 0x07-pointers_arrays_strings/main.h
@@ -12,7 +12,7 @@ char *_strpbrk(char *s, char *accept);
 
 char *_strstr(char *haystack, char *needle);
 
-int is_equal(char *h, char *n, unsigned int i, unsigned int j);
+int is_equal(char *h, char *n, long int i, long int j);
 
 void print_chessboard(char (*a)[8]);
